putunsigned, putptr added. ft_printf now return -1 if something wrong

+++ ft_printf/libft/ft_split.c
@@ -6,7 +6,7 @@
 /*   By: mukaplan <mukaplan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/16 12:37:01 by mukaplan          #+#    #+#             */
-/*   Updated: 2024/11/18 14:00:33 by mukaplan         ###   ########.fr       */
+/*   Updated: 2025/01/18 18:00:14 by mukaplan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,7 +57,7 @@ static char	*ft_new_word(char **cursor, char c)
 char	**ft_split(char const *s, char c)
 {
 	char			**strings;
-	unsigned int	i;
+	int	i;
 	unsigned int	word_count;
 	char			*cursor;
 
@@ -68,7 +68,7 @@ char	**ft_split(char const *s, char c)
 	if (strings == NULL)
 		return (NULL);
 	strings[word_count] = NULL;
-	while (i < word_count)
+	while (i < (int)word_count)
 	{
 		strings[i] = ft_new_word(&cursor, c);
 		if (strings[i] == NULL)
